Move XCB utilty functions to xcb_util.cpp
C++fy send_xcb_message()
